# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Versioned commerce API for the storefront and operations console
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/products:
    get:
      operationId: listProducts
      tags: [storefront]
      summary: List active products
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Category"
      responses:
        "200":
          description: Active storefront products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /v1/products/{productId}:
    get:
      operationId: getProduct
      tags: [storefront]
      summary: Get a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/registrations/eligibility:
    post:
      operationId: checkRegistrationEligibility
      tags: [storefront]
      summary: Check first-time registration eligibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationEligibilityInput"
      responses:
        "200":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationEligibility"
  /v1/registrations/claim:
    post:
      operationId: claimRegistrationOffer
      tags: [storefront]
      summary: Claim the first-time registration offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationEligibilityInput"
      responses:
        "201":
          description: Offer claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationClaim"
  /v1/discounts/validate:
    post:
      operationId: validateDiscount
      tags: [storefront]
      summary: Validate a dynamic discount code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountValidationInput"
      responses:
        "200":
          description: Discount validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscountValidation"
...
  /v1/admin/staff/login:
    post:
      operationId: staffLogin
      tags: [admin]
      summary: Staff credential login (issues a staff bearer token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /v1/admin/staff:
    get:
      operationId: listStaff
      tags: [admin]
      summary: List staff accounts
      security:
        - clerkSession: []
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
    post:
      operationId: createStaff
      tags: [admin]
      summary: Create a staff account
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                role: { type: string }
                expiresAtHours: { type: number }
                expiresAtDate: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /v1/admin/staff/{id}:
    put:
      operationId: updateStaff
      tags: [admin]
      summary: Update a staff account
      security:
        - clerkSession: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    delete:
      operationId: deleteStaff
      tags: [admin]
      summary: Delete a staff account
      security:
        - clerkSession: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { type: object, additionalProperties: true }

  # ─── Product approvals & storage (Clerk-protected) ───────────────────────
...
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: Clerk
    customerSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
    staffSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
...
    Search:
      name: search
      in: query
      schema:
        type: string
    Category:
      name: category
      in: query
      schema:
        type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, description, priceCents, category, imageUrl, status, featured, inventory, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        priceCents: { type: number, minimum: 0 }
        compareAtPriceCents: { type: ["number", "null"], minimum: 0 }
        category: { type: string }
        imageUrl: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        featured: { type: boolean }
        inventory: { type: number, minimum: 0 }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    RegistrationEligibilityInput:
      type: object
      required: [email]
      properties:
        email: { type: string, minLength: 3 }
    RegistrationEligibility:
      type: object
      required: [eligible, reason, offerCode]
      properties:
        eligible: { type: boolean }
        reason: { type: string }
        offerCode: { type: ["string", "null"] }
    RegistrationClaim:
      type: object
      required: [claimed, offerCode, message]
      properties:
        claimed: { type: boolean }
        offerCode: { type: ["string", "null"] }
        message: { type: string }
    DiscountValidationInput:
      type: object
      required: [code, subtotalCents, isFirstOrder]
      properties:
        code: { type: string }
        subtotalCents: { type: number, minimum: 0 }
        isFirstOrder: { type: boolean }
    DiscountValidation:
      type: object
      required: [valid, code, discountCents, message]
      properties:
        valid: { type: boolean }
        code: { type: string }
        discountCents: { type: number, minimum: 0 }
        message: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }